// src/shared/types/generated/application/ItemQuery.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ItemFilters } from "./ItemFilters";
import type { ItemScope } from "./ItemScope";
import type { ItemSort } from "./ItemSort";

export type ItemQuery = { scope: ItemScope, filters: ItemFilters, sort: ItemSort, };